🐛 fix: lift content state from ReadmeResult to ResultTemplate
After refactoring to the command bus system, it is no longer possible to register multiple handlers for the same command. The content state was lifted to ResultTemplate, allowing it to be shared without additional command handlers. As a result, CopyCurrentFileContent was removed.
